requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-ranged-response==0.2.0
django-recaptcha==3.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
